+++ nexus/src/app/mod.rs
@@ -63,6 +63,7 @@ mod instance;
 mod instance_network;
 mod internet_gateway;
 mod ip_pool;
+mod login;
 mod metrics;
 mod network_interface;
 pub(crate) mod oximeter;
